cifs: channel helper functions should set tcpStatus
When specific channels of a session (or all it's channels)
are identified for reconnect, we set the correct bits
on the chans_need_reconnect bitmask. However, we also
need to set the tcpStatus to CifsNeedReconnect. This
will tell the cifsd thread to trigger reconnect.
Change the helper functions to do this.

Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>