cifs: use session_estab from primary channel
We use the session_estab field in the server struct to
identify whether a session is setup, so that the requests
can be signed appropriately.

With multichannel, we used to mark this field for each
channel during reconnect. This is wrong, since we have
an active session, unless all the channels are down.

With this fix, session_estab is always referenced in
the server struct for the primary channel.

Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>