cifs: force close all open handles during umount
With deferred close for files, we can keep handles open,
with references to dentry and super block, even after umount.
This results in inability to remove cifs.ko module, even after
umount; since the cifs data structures are still allocated.

With this change, we drop all references to VFS objects in
cifs_close, and take them again if a deferred handle is selected
in cifs_open. Due to this fixed refcounting, cifs_kill_sb gets
called at the time of umount, and we force closing all
deferred handles there.

Fixes: c3f207ab29f7 ("cifs: Deferred close for files")
Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>