cifs: avoid reconnect race in setting session and tcon status
Reconnect is a complex operation in cifs, with cifsd and other
threads combining to bring about reconnect. cifs_reconnect,
which is always called from cifsd should only mark connections,
session and tcon as needing reconnet. The actual task of reconnect
happens in i/o threads by calling smb2_reconnect. This is where
the status fields of these structs should be set.

We had cifsd setting the status, and hence resulting in weird
race conditions. This change moves the task of setting the status
to cifs_setup_session and cifs_tree_connect (which are called by
smb2_reconnect).

This change also ensures that the session and tcon status fields
are changed in such a way that several simultaneous reconnects
do not leave them unusable.

Cc: <Stable@vger.kernel.org>
Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>