cifs: maintain an open list of directories too
Currently, we do not maintain an open directory handle
list. Without handle caching, which is only done when
dir lease is supported, it is not possible to iterate
over open directories now.

This patch introduces a new list openDirList in tcon.
This list is appended to when a local handle is created
for dirs.

/proc/fs/cifs/open_dirs now lists these local dir handles.
These should ideally list remote handles too, which they
don't do today. Will cover that in a future patch.

Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>